+++ src/main/java/br/com/itau/grupo4/ticketsmicroservice/adapter/client/payment/dto/RefundRequest.java
@@ -1,4 +1,4 @@
-package br.com.itau.grupo4.ticketsmicroservice.client.payment.dto;
+package br.com.itau.grupo4.ticketsmicroservice.adapter.client.payment.dto;
 
 import br.com.itau.grupo4.ticketsmicroservice.enums.TicketStatus;
 import br.com.itau.grupo4.ticketsmicroservice.enums.TicketType;
