+++ src/main/java/br/com/itau/grupo4/ticketsmicroservice/adapter/client/payment/service/PaymentService.java
@@ -1,8 +1,7 @@
-package br.com.itau.grupo4.ticketsmicroservice.client.payment.service;
+package br.com.itau.grupo4.ticketsmicroservice.adapter.client.payment.service;
 
-import br.com.itau.grupo4.ticketsmicroservice.client.payment.dto.RefundRequest;
-import br.com.itau.grupo4.ticketsmicroservice.client.payment.dto.RefundResponse;
-import br.com.itau.grupo4.ticketsmicroservice.client.session.dto.SessionRequest;
+import br.com.itau.grupo4.ticketsmicroservice.adapter.client.payment.dto.RefundResponse;
+import br.com.itau.grupo4.ticketsmicroservice.adapter.client.payment.dto.RefundRequest;
 import org.springframework.stereotype.Component;
 import org.springframework.web.reactive.function.client.WebClient;
 import reactor.core.publisher.Mono;
