+++ src/main/java/br/com/itau/grupo4/ticketsmicroservice/service/TicketService.java
@@ -1,10 +1,10 @@
 package br.com.itau.grupo4.ticketsmicroservice.service;
 
-import br.com.itau.grupo4.ticketsmicroservice.adapter.qrcodeapi.GenerateQrCodeAPI;
-import br.com.itau.grupo4.ticketsmicroservice.client.payment.dto.RefundRequest;
-import br.com.itau.grupo4.ticketsmicroservice.client.payment.service.PaymentService;
-import br.com.itau.grupo4.ticketsmicroservice.client.session.dto.SessionRequest;
-import br.com.itau.grupo4.ticketsmicroservice.client.session.service.SessionService;
+import br.com.itau.grupo4.ticketsmicroservice.adapter.client.qrcodeapi.GenerateQrCodeAPI;
+import br.com.itau.grupo4.ticketsmicroservice.adapter.client.payment.dto.RefundRequest;
+import br.com.itau.grupo4.ticketsmicroservice.adapter.client.payment.service.PaymentService;
+import br.com.itau.grupo4.ticketsmicroservice.adapter.client.session.dto.SessionRequest;
+import br.com.itau.grupo4.ticketsmicroservice.adapter.client.session.service.SessionService;
 import br.com.itau.grupo4.ticketsmicroservice.dto.CanceledTicketResponse;
 import br.com.itau.grupo4.ticketsmicroservice.dto.TicketResponse;
 import br.com.itau.grupo4.ticketsmicroservice.exception.TicketNotFoundException;
